requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
brotli==1.0.9
    # via flask-compress
click==8.1.3
    # via flask
flask==2.1.3
    # via
    #   -r requirements.in
    #   flask-compress
flask-compress==1.12
    # via -r requirements.in
flask-talisman==1.0.0
    # via -r requirements.in
govuk-frontend-jinja==2.3.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   flask
    #   govuk-frontend-jinja
markupsafe==2.1.1
    # via jinja2
python-dotenv==0.20.0
    # via -r requirements.in
werkzeug==2.1.2
    # via flask
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
attrs==21.4.0
bandit==1.7.4
black==22.6.0
bokeh==2.4.3
Brotli==1.0.9
build==0.8.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
deepmerge==1.0.1
dnspython==2.2.1
dparse==0.5.1
email-validator==1.2.1
flake8==3.8.4
flake8-bugbear==22.7.1
flake8-polyfill==1.0.2
Flask-Assets==2.0
Flask-WTF==1.0.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.13.0
google-auth-oauthlib==0.5.3
govuk-frontend-jinja==2.3.0
govuk-frontend-wtf==2.1.0
gspread==5.6.0
gspread-pandas==3.2.2
hacking==4.1.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
jsmin==3.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pathspec==0.9.0
pbr==5.9.0
pep517==0.12.0
pep8-naming==0.11.1
Pillow==9.2.0
pip-tools==6.8.0
piprot==0.9.11
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safety==2.1.1
six==1.16.0
smmap==5.0.0
stevedore==4.0.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.11
webassets==2.0
WTForms==3.0.1
